+++ src/lib/bindings/gsc/ParcelDelivery.spec.ts
@@ -1,9 +1,9 @@
 import { Integer, Primitive, Sequence } from 'asn1js';
 
-import { arrayBufferFrom, expectBuffersToEqual } from '../../../_test_utils';
-import { derSerializeHeterogeneousSequence } from '../../../asn1';
-import { derDeserialize } from '../../../crypto_wrappers/_utils';
-import InvalidMessageError from '../../InvalidMessageError';
+import { arrayBufferFrom, expectBuffersToEqual } from '../../_test_utils';
+import { derSerializeHeterogeneousSequence } from '../../asn1';
+import { derDeserialize } from '../../crypto_wrappers/_utils';
+import InvalidMessageError from '../../messages/InvalidMessageError';
 import { ParcelDelivery } from './ParcelDelivery';
 
 const DELIVERY_ID = 'the id';
