+++ src/lib/bindings/gsc/ParcelDelivery.ts
@@ -1,8 +1,8 @@
 import { OctetString, Primitive, verifySchema, VisibleString } from 'asn1js';
 import { TextDecoder } from 'util';
 
-import { derSerializeHeterogeneousSequence, makeHeterogeneousSequenceSchema } from '../../../asn1';
-import InvalidMessageError from '../../InvalidMessageError';
+import { derSerializeHeterogeneousSequence, makeHeterogeneousSequenceSchema } from '../../asn1';
+import InvalidMessageError from '../../messages/InvalidMessageError';
 
 export class ParcelDelivery {
   public static deserialize(serialization: ArrayBuffer): ParcelDelivery {
