+++ src/lib/bindings/gsc/PrivateNodeRegistrationAuthorization.spec.ts
@@ -1,13 +1,13 @@
 import { ObjectIdentifier, OctetString, Primitive, Sequence, VisibleString } from 'asn1js';
 import moment from 'moment';
-import { generateRSAKeyPair } from '../../../..';
+import { generateRSAKeyPair } from '../../../index';
 
-import { arrayBufferFrom } from '../../../_test_utils';
-import { dateToASN1DateTimeInUTC, derSerializeHeterogeneousSequence } from '../../../asn1';
-import { derDeserialize } from '../../../crypto_wrappers/_utils';
-import { verify } from '../../../crypto_wrappers/rsaSigning';
-import { RELAYNET_OIDS } from '../../../oids';
-import InvalidMessageError from '../../InvalidMessageError';
+import { arrayBufferFrom } from '../../_test_utils';
+import { dateToASN1DateTimeInUTC, derSerializeHeterogeneousSequence } from '../../asn1';
+import { derDeserialize } from '../../crypto_wrappers/_utils';
+import { verify } from '../../crypto_wrappers/rsaSigning';
+import { RELAYNET_OIDS } from '../../oids';
+import InvalidMessageError from '../../messages/InvalidMessageError';
 import { PrivateNodeRegistrationAuthorization } from './PrivateNodeRegistrationAuthorization';
 
 describe('PrivateNodeRegistrationAuthorization', () => {
